Naver_Flights/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pip==24.2
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.32.3
setuptools==72.1.0
urllib3==2.2.2
wheel==0.43.0
